[opt]: Narrow data type of parts from int32 to int8

+++ tests/cuda_ext/test_ari_py.cu
@@ -11,16 +11,16 @@ int main() {
 
     try {
         // Define vectors in C++
-        std::vector<int> part0 = {2, 3, 6, 1, 0, 5, 4, 3, 6, 2};
-        std::vector<int> part1 = {0, 6, 2, 5, 1, 3, 4, 6, 0, 2};
+        std::vector<int8_t> part0 = {2, 3, 6, 1, 0, 5, 4, 3, 6, 2};
+        std::vector<int8_t> part1 = {0, 6, 2, 5, 1, 3, 4, 6, 0, 2};
 
         // Import required Python modules
         py::module_ np = py::module_::import("numpy");
         py::module_ ccc_module = py::module_::import("ccc.sklearn.metrics");
 
         // Convert C++ vectors to numpy arrays
-        py::array_t<int> np_part0 = py::cast(part0);
-        py::array_t<int> np_part1 = py::cast(part1);
+        py::array_t<int8_t> np_part0 = py::cast(part0);
+        py::array_t<int8_t> np_part1 = py::cast(part1);
 
         // Call the ccc function
         py::object result = ccc_module.attr("adjusted_rand_index")(np_part0, np_part1);
